ai-models/requirements.txt
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
tensorflow==2.13.0
keras==2.13.1
xgboost==1.7.6
lightgbm==4.0.0
joblib==1.3.2
requests==2.31.0
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
pymongo==4.5.0
schedule==1.2.0
geopy==2.4.0
meteostat==1.6.5
